server/requirements.txt
django
django-cors-headers
djangorestframework
markdown
django-filter
pandas
tqdm
requests
retry